+++ cognite/experimental/data_classes/contextualization.py
@@ -99,9 +99,9 @@ class ContextualizationJob(CogniteResource):
             self.error_message,
         )
 
-    @staticmethod
-    def _load_with_status(data, status_path, cognite_client):
-        obj = ContextualizationJob._load(data, cognite_client=cognite_client)
+    @classmethod
+    def _load_with_status(cls, data, status_path, cognite_client):
+        obj = cls._load(data, cognite_client=cognite_client)
         obj._status_path = status_path
         return obj
 
@@ -246,6 +246,29 @@ class EntityMatchingModelList(CogniteResourceList):
     _UPDATE = EntityMatchingModelUpdate
 
 
+class EntityMatchingPipelineRun(ContextualizationJob):
+    def __init__(self, pipeline_id=None, **kwargs):
+        super().__init__(**kwargs)
+        self.pipeline_id = pipeline_id
+        self._status_path = "/context/entitymatching/pipelines/run/"  # since we can list this, would like .result even if we didn't this via .run
+
+    @property
+    def suggested_rules(self):
+        """List of suggested rules. Depends on .result and may block"""
+        return self.result["suggestedRules"]
+
+    @property
+    def matches(self):
+        """List of matches. Depends on .result and may block"""
+        return self.result["matches"]
+
+
+class EntityMatchingPipelineRunList(CogniteResourceList):
+    _RESOURCE = EntityMatchingPipelineRun
+    _UPDATE = None
+    _ASSERT_CLASSES = False
+
+
 class EntityMatchingPipeline(CogniteResource):
     _RESOURCE_PATH = "/context/entitymatching/pipelines"
     _STATUS_PATH = _RESOURCE_PATH + "/"
@@ -261,7 +284,9 @@ class EntityMatchingPipeline(CogniteResource):
         targets: Dict = None,
         true_matches: List = None,
         rejected_matches: List = None,
+        confirmed_matches: List = None,
         score_threshold: float = None,
+        schedule_interval: int = None,
         rules: List = None,
         status=None,
         error_message=None,
@@ -274,10 +299,13 @@ class EntityMatchingPipeline(CogniteResource):
         The fields below can be filled when creating a pipeline. Other fields should be left empty, and return status information on successful creation and retrieval.
         Args:
             external_id, name, description: standard fields for a resource.
-            model_parameters: TODO: doc
-            sources, targets: TODO: doc
-            true_matches: true matches to use in training and overriding any other results
+            model_parameters: a dictionary with fields `match_fields`, `feature_type`, `classifier`, as in the `fit` method for entity matching.
+            sources, targets: a dictionary of the format {'resource': ..., 'dataSetIds': [{'id':...},{'externalId':...}]}
+            true_matches: existing matches with reasonable certainty to use in training.
+            confirmed_matches: user-confirmed certain matches which will be used to override any other results.
+            rejected_matches: user-confirmed wrong results which will be used to blank output for a match result if it is one of these.
             rules: list of matching rules
+            schedule_interval: automatically schedule pipeline to be run every this many seconds.
         """
 
         self.id = id
@@ -288,9 +316,11 @@ class EntityMatchingPipeline(CogniteResource):
         self.sources = sources
         self.targets = targets
         self.true_matches = true_matches
+        self.confirmed_matches = confirmed_matches
         self.rejected_matches = rejected_matches
         self.score_threshold = score_threshold
         self.rules = rules
+        self.schedule_interval = schedule_interval
 
         self.status = status
         self.created_time = created_time
@@ -300,28 +330,73 @@ class EntityMatchingPipeline(CogniteResource):
 
         self._cognite_client = cognite_client
 
-    def run(self):
-        self._cognite_client.entity_matching.pipelines.run(id=self.id)
+    def run(self) -> EntityMatchingPipelineRun:
+        return self._cognite_client.entity_matching.pipelines.run(id=self.id)
+
+    def runs(self) -> EntityMatchingPipelineRunList:
+        return self._cognite_client.entity_matching.pipelines.runs.list(id=self.id)
+
+    def latest_run(self) -> EntityMatchingPipelineRun:
+        return self._cognite_client.entity_matching.pipelines.runs.retrieve_latest(id=self.id)
 
 
 class EntityMatchingPipelineUpdate(CogniteUpdate):  # not implemented yet
-    pass
+    """Changes applied to entity matching pipeline
+
+    Args:
+        id (int): A server-generated ID for the object.
+        external_id (str): The external ID provided by the client. Must be unique for the resource type.
+    """
+
+    class _PrimitiveUpdate(CognitePrimitiveUpdate):
+        def set(self, value: Any) -> "EntityMatchingPipelineUpdate":
+            return self._set(value)
+
+    @property
+    def name(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "name")
+
+    @property
+    def description(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "description")
+
+    @property
+    def model_parameters(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "modelParameters")
+
+    @property
+    def sources(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "sources")
+
+    @property
+    def targets(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "targets")
+
+    @property
+    def true_matches(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "trueMatches")
+
+    @property
+    def confirmed_matches(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "confirmedMatches")
+
+    @property
+    def rejected_matches(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "rejectedMatches")
+
+    @property
+    def rules(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "rules")
+
+    @property
+    def score_threshold(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "scoreThreshold")
+
+    @property
+    def schedule_interval(self):
+        return EntityMatchingPipelineUpdate._PrimitiveUpdate(self, "scheduleInterval")
 
 
 class EntityMatchingPipelineList(CogniteResourceList):
     _RESOURCE = EntityMatchingPipeline
     _UPDATE = EntityMatchingPipelineUpdate
-
-
-class EntityMatchingPipelineRun(ContextualizationJob):
-    _COMMON_FIELDS = ContextualizationJob._COMMON_FIELDS & {"pipeline_id"}
-
-    def __init__(self, pipeline_id=None, **kwargs):
-        super().__init__(**kwargs)
-        self.pipeline_id = pipeline_id
-
-
-class EntityMatchingPipelineRunList(CogniteResourceList):
-    _RESOURCE = EntityMatchingPipelineRun
-    _UPDATE = None
-    _ASSERT_CLASSES = False